refactor: script 컴포넌트 분리
