+++ src/interfaces/interfaces.ts
@@ -17,3 +17,14 @@ export interface LocationType {
   url: string;
   script: Array<{ start: string; end: string; text: string }>;
 }
+
+export interface ScriptType {
+  script: {
+    start: string;
+    end: string;
+    text: string;
+  };
+  currentTime: number | undefined;
+  handleTimeline: (e: React.MouseEvent<HTMLParagraphElement>) => void;
+  convertTime: (arr: Array<string>) => number;
+}
